exclude the failing tests only for `replay-netty4`
